+++ delta-coverage-gradle/src/test/kotlin/io/github/surpsg/deltacoverage/gradle/sources/SourcesResolverTest.kt
@@ -1,6 +1,6 @@
 package io.github.surpsg.deltacoverage.gradle.sources
 
-import io.github.surpsg.deltacoverage.CoverageEngine
+import io.github.surpsg.deltacoverage.gradle.CoverageEngine
 import io.github.surpsg.deltacoverage.gradle.unittest.testJavaProject
 import io.kotest.assertions.assertSoftly
 import io.kotest.assertions.throwables.shouldThrow